Rework appraisals to match modern BC apps
This removes the Rails 3.x version testing in favor of 4.x and 5.x
testing. Now, with 2.x, 4.x, and 5.x tested, it better reflects the
landscape of the BC apps. The appraisals now use a `rails` prefix to
make it more clear what the number means in the gemfile.

Also, the 3.x appraisals simply didn't work or install on my machine.

As part of this change, I updated the Travis CI config to hopefully
work.